+++ SnipeitPS/Public/Set-SnipeitComponent.ps1
@@ -17,6 +17,9 @@ ID number of category
 .PARAMETER qty
 Quantity of the components you have
 
+.PARAMETER min_amt
+Minimum Quantity of the components before alert is triggered
+
 .PARAMETER location_id
 ID number of the location the accessory is assigned to
 
@@ -55,11 +58,13 @@ function Set-SnipeitComponent()
         [parameter(mandatory = $true)]
         [int]$qty,
 
+        [Nullable[System.Int32]]$min_amt,
+
         [string]$name,
 
-        [int]$company_id,
+        [Nullable[System.Int32]]$company_id,
 
-        [int]$location_id,
+        [Nullable[System.Int32]]$location_id,
 
 
         [string]$order_number,